+++ docker-compose.prod.yml
@@ -28,6 +28,7 @@ services:
   backend-blue:
     image: ${DOCKER_USER}/tab-backend:blue
     container_name: backend-blue
+    restart: always
     expose:
       - 8080
     ports:
@@ -43,6 +44,7 @@ services:
   backend-green:
     image: ${DOCKER_USER}/tab-backend:green
     container_name: backend-green
+    restart: always
     expose:
       - 8080
     ports:
