+++ docker-compose.yml
@@ -17,31 +17,9 @@ services:
     volumes:
       - tab-backend:/usr/src/mysql
 
-#  backend:
-#    build: .
-#    container_name: backend
-#    expose:
-#      - 8080
-#    ports:
-#      - "8080:8080"
-#    environment:
-#      SPRING_DATASOURCE_URL: jdbc:mysql://mysql:3306/tab
-#      SPRING_DATASOURCE_USER: root
-#      SPRING_DATASOURCE_PASSWORD: ${DB_PASSWORD}
-#    depends_on:
-#      - mysql
-
-  redis:
-    image: redis:latest
-    container_name: redis
-    ports:
-      - "6379:6379"
-    volumes:
-      - tab-backend:/usr/src/redis
-
-  backend-blue:
+  backend:
     build: .
-    container_name: backend-blue
+    container_name: backend
     expose:
       - 8080
     ports:
@@ -54,32 +32,13 @@ services:
       - mysql
       - redis
 
-  backend-green:
-    build: .
-    container_name: backend-green
-    expose:
-      - 8080
+  redis:
+    image: redis:latest
+    container_name: redis
     ports:
-      - "8081:8080"
-    environment:
-      SPRING_DATASOURCE_URL: jdbc:mysql://mysql:3306/tab
-      SPRING_DATASOURCE_USER: root
-      SPRING_DATASOURCE_PASSWORD: ${DB_PASSWORD}
-    depends_on:
-      - mysql
-      - redis
-
-  nginx:
-    image: nginx:latest
-    container_name: nginx
-    ports:
-      - "80:80"
-      - "443:443"
+      - "6379:6379"
     volumes:
-      - ./nginx/nginx.conf:/etc/nginx/nginx.conf
-    depends_on:
-      - backend-blue
-      - backend-green
+      - tab-backend:/usr/src/redis
 
 volumes:
   tab-backend:
