+++ src/main/resources/application.yml
@@ -1,0 +1,34 @@
+spring:
+  datasource:
+    include:
+      - docker
+  #      - local
+
+  jpa:
+    hibernate:
+      ddl-auto : validate  # 스테이징, 운영 환경에서는 create, create-drop, update는 사용하지 말것
+    properties:
+      hibernate:
+        format_sql: true
+        use_sql_comments: true    # 실행되는 JPQL을 보고싶다면 다음과 같이 추가한다.
+    show_sql: true
+
+logging:
+  level:
+    org:
+      hibernate: info
+
+springdoc:
+  swagger-ui:
+    path: /swagger-ui.html
+    groups-order: DESC
+    operations-sorter: method   # 컨트롤러를 원하는 순서대로 정렬할 수 있다.
+    disable-swagger-default-url: true
+    display-request-duration: true
+  api-docs:
+    path: /api-docs
+  show-actuator: true
+  default-consumes-media-type: application/json
+  default-produces-media-type: application/json
+  paths-to-match:   # 보여주고 싶은 컨트롤러만 보여주기
+    - /api/v1/**
